# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Coziyoo API v1
  version: 1.0.0
  description: Versioned API contract for Coziyoo marketplace backend.
servers:
  - url: http://localhost:3000
    description: Local
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /v1/sellers/{sellerId}/finance/reports:
    post:
      tags: [finance]
      parameters:
        - in: path
          name: sellerId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: Seller report generated }
    get:
      tags: [finance]
      parameters:
        - in: path
          name: sellerId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Seller reports list }
...